Polishing - pull methods up to abstract factory
